Move simple Fast Path API up to `RelationalProfile`
The old `JdbcFastPath` did not actually require JDBC because it relays
all calls to the specialized ResultConverters that it wraps. It is now
called `SimpleFastPathResultConverter` and it still requires a
`TypeMappingResultConverter` on top of a `ProductResultConverter` to
work but this has always been enforced at runtime (falling back to the
non-FastPath implementation in other cases) and is applicable outside of
`JdbcProfile` as well.
